requirements.txt

python-etcd==0.4.5
PyYAML==5.1.2